LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.get.unaffected-expected.txt

2d.imageData.get.unaffected
getImageData() is not affected by context state
Actual output:
Expected output:


PASS getImageData() is not affected by context state
